Passthrough is working.
git-svn-id: http://chianti.ucsd.edu/svn@22104 0ecc0d97-ab19-0410-9704-bfe1a75892f5

+++ core3/ding-presentation-impl/branches/vp-tree/src/main/java/org/cytoscape/ding/NodeView.java
@@ -106,12 +106,12 @@ public interface NodeView  extends GraphViewObject {
   /**
    * @param trans new value for the transparency
    */
-  public void setTransparency ( float trans );
+  public void setTransparency ( int trans );
   
   /**
    * @return the value for the transparency for this node
    */
-  public float getTransparency ();
+  public int getTransparency ();
 
 
   /**
